requirements.txt
pyasn1==0.6.0
python-dotenv==1.0.1
rsa==4.9
